Subscribe to TaskScheduler.UnobservedTaskException
Tasks that do not handle exceptions should also be reported as UnhandledException.
The MauiExceptions.cs file did not have this, but it is a valuable exception to report
as typically there would be no in-app indication/crash if any exceptions were thrown
in Tasks that are not await-ed or Wait()-ed.

The original has some comments:
https://gist.github.com/mattjohnsonpint/7b385b7a2da7059c4a16562bc5ddb3b7?permalink_comment_id=4710633#gistcomment-4710633

This code is based on the updated gist: https://gist.github.com/myrup/43ee8038e0fd6ef4d31cbdd67449a997#file-mauiexceptions-cs-L21-L25