Recompiled iOS network monitor for arm64 simulator
The library was previously compiled for iPhoneOS arm64, and iPhoneSimulator x64 as it was compiled before Apple Silicon. This would prevent the project form being compiled as it would try linking the arm64 iPhoneSimulator to the x64 library.

The library has now been compiled for iPhoneOS arm64, and iPhoneSimulator arm64/x64.