Created Raygun4Maui.Platform project
This project will contain the Android and iOS/MacCatalyst (and possibly other binding libraries in the future). The Raygun4Maui SDK will now depend on the NuGet reference of this newly created library meaning we can build the project on Windows again as we do not compile the Objective-C binding library each time now.